+++ lib/src/destination.dart
@@ -25,10 +25,12 @@ class Destination {
   });
 
   /// Create a [Destination] from [RouteInformation].
-  factory Destination.fromRouteInformation(RouteInformation routeInformation,) {
+  factory Destination.fromRouteInformation(
+    RouteInformation routeInformation,
+  ) {
     final Object? argument = routeInformation.state;
     return Destination(
-      path: routeInformation.location ?? '' /*TODO think how to handle null*/,
+      path: routeInformation.location,
       metadata: argument is DestinationMetadata
           ? argument
           : DestinationMetadata(arguments: argument),
@@ -61,10 +63,10 @@ class Destination {
   @override
   bool operator ==(Object other) =>
       identical(this, other) ||
-          other is Destination &&
-              runtimeType == other.runtimeType &&
-              path == other.path &&
-              metadata == other.metadata;
+      other is Destination &&
+          runtimeType == other.runtimeType &&
+          path == other.path &&
+          metadata == other.metadata;
 
   @override
   int get hashCode => path.hashCode ^ metadata.hashCode;
@@ -134,10 +136,10 @@ class DestinationMetadata {
   @override
   bool operator ==(Object other) =>
       identical(this, other) ||
-          other is DestinationMetadata &&
-              runtimeType == other.runtimeType &&
-              arguments == other.arguments &&
-              listEquals<Destination>(history, other.history);
+      other is DestinationMetadata &&
+          runtimeType == other.runtimeType &&
+          arguments == other.arguments &&
+          listEquals<Destination>(history, other.history);
 
   @override
   int get hashCode => arguments.hashCode ^ history.hashCode;