+++ lib/src/utils/db_router_delegate_extensions.dart
@@ -20,9 +20,13 @@ extension DBRouterDelegateBuildContextExtensions on BuildContext {
   Future<T?> navigateTo<T>({
     required String location,
     Object? arguments,
+    List<Destination>? history,
     bool useRootDelegate = false,
   }) {
-    return DBRouterDelegate.of(this, root: useRootDelegate)
-        .navigateTo<T>(location: location, arguments: arguments);
+    return DBRouterDelegate.of(this, root: useRootDelegate).navigateTo<T>(
+      location: location,
+      arguments: arguments,
+      history: history,
+    );
   }
 }